requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dateutil==2.8.2
